+++ CHANGELOG.md
@@ -1,8 +1,30 @@
 ## v2.1.0
 
--   Add new `find(iterable, predicate)` function, which is almost the same as
-    `first()` but behaves slightly more intuitive in the case where no
-    predicate function is given.
+The following functions retain richer type information about their arguments:
+
+-   `ifilter()`
+-   `filter()`
+-   `partition()`
+
+For example, TypeScript will now know the following:
+
+```ts
+const items = [3, 'hi', -7, 'foo', 13];
+
+function isNum(value: unknown): value is number {
+    return typeof value === 'number';
+}
+
+const numbers: number[] = filter(items, isNum); // ✅
+
+const [numbers, strings] = partition(items, isNum); // ✅
+//     ^^^^^^^  ^^^^^^^ string[]
+//     number[]
+```
+
+-   Add new `find(iterable, pred)` function, which is almost the same as
+    `first(iterable, pred)` but behaves slightly more intuitive in the case
+    where no predicate function is given.
 
 ## v2.0.0